requirements.txt
python-telegram-bot==21.6
openpyxl>=3.1.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0
flask>=2.3.0
gunicorn>=21.2.0
pytz
